# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Brand Service API",
    "description": "Microservice for managing brand information, media assets, organization data, and AI-powered content analysis.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3005"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
  "components": {
    "schemas": {
...
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "code",
          "field",
          "message"
        ]
      },
...
      "SetBrandWebsiteResponse": {
        "type": "object",
        "properties": {
          "brandId": {
            "type": "string"
          },
          "domain": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "brandId",
          "domain",
          "name",
          "url"
        ]
      },
      "DomainConflictErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "enum": [
              "DOMAIN_OWNED_BY_YOUR_PAID_BRAND",
              "DOMAIN_OWNED_BY_ANOTHER_ORG"
            ]
          },
          "message": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "conflictingBrandId": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "code",
          "message",
          "domain",
          "conflictingBrandId"
        ],
        "description": "A domain can only be refused when somebody has CHECKED OUT on the brand holding it. `code` tells the two refusals apart so a UI can render distinct copy: DOMAIN_OWNED_BY_YOUR_PAID_BRAND (the caller's own org already paid on another brand with this domain — point the user at that brand) vs DOMAIN_OWNED_BY_ANOTHER_ORG (a different, paying organization holds it). `conflictingBrandId` is the brand holding the domain.",
        "example": {
          "error": "Your organization already has a paid brand on domain \"acme.com\"",
          "code": "DOMAIN_OWNED_BY_YOUR_PAID_BRAND",
          "message": "Your organization already has a paid brand on domain \"acme.com\"",
          "domain": "acme.com",
          "conflictingBrandId": "0b5f5b1e-2b1a-4a6a-9f1e-0d3a5c7b9e11"
        }
      },
      "SetBrandWebsiteRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "description": "Attach a website to an existing brand (e.g. a no-website brand whose user later adds their site). URL may be a bare domain or full URL; normalized server-side. Sets the brand's url + domain. The next post-cache-expiry field extraction re-sources from the site (rides the existing field cache — no new TTL).",
        "example": {
          "url": "https://acme.com"
        }
      },
...
      "GetBrandResponse": {
        "type": "object",
        "properties": {
          "brand": {
            "$ref": "#/components/schemas/BrandDetail"
          }
        },
        "required": [
          "brand"
        ]
      },
      "BrandDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Brand UUID"
          },
          "domain": {
            "type": "string",
            "nullable": true,
            "description": "Normalized domain (subdomains preserved, www stripped). `null` for a no-website brand (identified by `name`)."
          },
          "url": {
            "type": "string",
            "nullable": true,
            "description": "Full brand website URL. `null` for a no-website brand."
          },
          "name": {
            "type": "string",
            "description": "Brand display name. Lazy-extracted from the website on first read if missing; user-provided for a no-website brand."
          },
          "logoUrl": {
            "type": "string",
            "nullable": true,
            "description": "Logo image URL. Lazy-filled with a deterministic logo.dev URL on first read if missing. `null` for a no-website brand (no domain to build one from)."
          },
          "clickDestinationUrl": {
            "type": "string",
            "nullable": true,
            "description": "Page outreach clicks should land on. Defaults to the brand's own landing URL (`url`) when the user has not set an override. `null` only for a no-website brand with no override (no landing URL to fall back to). Per-brand config, set via PUT /orgs/brands/{brandId}/click-destination."
          },
          "whatsAppLink": {
            "type": "string",
            "nullable": true,
            "description": "The brand's WhatsApp link — the click destination for the \"maximize WhatsApp conversations\" goal. `null` when unset (no sensible default, unlike clickDestinationUrl). Per-brand config, set via PUT /orgs/brands/{brandId}/whatsapp-link."
          },
          "colors": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            },
            "description": "The brand's own colour palette — hex strings in the order logo.dev's Brand API returns them (e.g. [\"#000103\",\"#ce2e36\",\"#003366\"]). Nothing is pre-filtered or ranked; the consumer selects. `null` means WE HAVE NO COLOURS for this brand (the provider has not indexed the domain yet, or has no palette for it) — a first-class answer a consumer falls back to its own charter on. No colour is ever invented, defaulted, or derived from the logo, the name, or the domain."
          },
          "createdAt": {
            "type": "string",
            "description": "ISO timestamp when the brand row was created."
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO timestamp when the brand row was last updated."
          }
        },
        "required": [
          "id",
          "domain",
          "url",
          "name",
          "logoUrl",
          "clickDestinationUrl",
          "whatsAppLink",
          "colors",
          "createdAt",
          "updatedAt"
        ]
      },
...
    },
...
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Service-to-service API key"
      }
    }
  },
  "paths": {
...
    "/orgs/brands/{brandId}": {
      "patch": {
        "summary": "Attach a website to an existing brand",
        "description": "Sets brands.url + brands.domain on an existing brand (e.g. a no-website brand whose user later adds their site). The next post-cache-expiry field extraction re-sources from the site automatically (rides the existing field cache — no new TTL). A domain belongs to whoever has CHECKED OUT on it (client-service is the source of truth). If another brand already holds the derived domain but NOBODY ever checked out on it, the domain is moved onto this brand and the abandoned holder is left as a no-website brand (and, when it belongs to the caller's own org, its data is merged in and it is removed from the caller's brand list). Only a holder somebody paid for is a 409, with a distinct `code` per case: `DOMAIN_OWNED_BY_YOUR_PAID_BRAND` (the caller's own org already checked out on it) or `DOMAIN_OWNED_BY_ANOTHER_ORG`.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetBrandWebsiteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Website attached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetBrandWebsiteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid brand ID or URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Brand not owned by the requesting org"
          },
          "404": {
            "description": "Brand not found"
          },
          "409": {
            "description": "Domain is held by a brand somebody has checked out on. `code` distinguishes DOMAIN_OWNED_BY_YOUR_PAID_BRAND from DOMAIN_OWNED_BY_ANOTHER_ORG.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainConflictErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          },
          "502": {
            "description": "client-service could not answer whether the holding brand was checked out (code CHECKOUT_STATUS_UNAVAILABLE)"
          }
        }
      }
    },
...
    "/internal/brands/{id}": {
      "get": {
        "summary": "Get a single brand by ID",
        "description": "Returns the canonical minimal brand shape (identity + name + logoUrl). All business fields (industry, target audience, mission, etc.) must be fetched via POST /internal/brands/extract-fields. name is never null: it is normally resolved at brand-create; this route keeps the lazy-fill safety net for rows created before that (logo.dev company index → landing-page HTML → titlecased domain, no LLM / run / cost). logoUrl is lazy-filled from a deterministic logo.dev image URL when null.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "orgId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Brand details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBrandResponse"
                }
              }
            }
          },
          "404": {
            "description": "Brand not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/public/brands/{id}": {
      "get": {
        "summary": "Get a single brand by ID (public, no auth)",
        "description": "Public mirror of GET /internal/brands/{id}. Identical response shape — identity + lazy-filled name and logoUrl. Use this when no API key is available (dashboards, embeddable widgets). Business fields must still be fetched via POST /orgs/brands/extract-fields (org auth required).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "orgId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Brand details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBrandResponse"
                }
              }
            }
          },
          "404": {
            "description": "Brand not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}